Make a new types module and add BinaryChange to get uint8array
