+++ automerge-backend-wasm/src/lib.rs
@@ -1,16 +1,17 @@
 //#![feature(set_stdio)]
 
 mod ser;
+mod types;
 
 use automerge_backend::{AutomergeError, Backend, Change};
-use automerge_backend::{BloomFilter, SyncHave, SyncMessage, SyncState};
+use automerge_backend::{SyncMessage, SyncState};
 use automerge_protocol::{ChangeHash, UncompressedChange};
 use js_sys::{Array, Uint8Array};
 use serde::de::DeserializeOwned;
-use serde::Deserialize;
 use serde::Serialize;
 use std::convert::TryFrom;
 use std::fmt::Display;
+use types::{RawSyncMessage, RawSyncState};
 use wasm_bindgen::prelude::*;
 use wasm_bindgen::JsCast;
 
@@ -198,163 +199,6 @@ fn export_changes(changes: Vec<&Change>) -> Array {
         result.push(bytes.as_ref());
     }
     result
-}
-
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct RawSyncState {
-    shared_heads: Vec<ChangeHash>,
-    last_sent_heads: Option<Vec<ChangeHash>>,
-    their_heads: Option<Vec<ChangeHash>>,
-    their_need: Option<Vec<ChangeHash>>,
-    our_need: Vec<ChangeHash>,
-    their_have: Option<Vec<RawSyncHave>>,
-    sent_changes: Vec<Vec<u8>>,
-}
-
-impl TryFrom<SyncState> for RawSyncState {
-    type Error = AutomergeError;
-
-    fn try_from(value: SyncState) -> Result<Self, Self::Error> {
-        let have = if let Some(have) = value.their_have {
-            Some(
-                have.into_iter()
-                    .map(RawSyncHave::try_from)
-                    .collect::<Result<_, _>>()?,
-            )
-        } else {
-            None
-        };
-        let sent_changes = value
-            .sent_changes
-            .into_iter()
-            .map(|c| c.raw_bytes().to_vec())
-            .collect();
-        Ok(Self {
-            shared_heads: value.shared_heads,
-            last_sent_heads: value.last_sent_heads,
-            their_heads: value.their_heads,
-            their_need: value.their_need,
-            our_need: value.our_need,
-            their_have: have,
-            sent_changes,
-        })
-    }
-}
-
-impl TryFrom<RawSyncState> for SyncState {
-    type Error = AutomergeError;
-
-    fn try_from(value: RawSyncState) -> Result<Self, Self::Error> {
-        let have = if let Some(have) = value.their_have {
-            Some(
-                have.into_iter()
-                    .map(SyncHave::try_from)
-                    .collect::<Result<_, _>>()?,
-            )
-        } else {
-            None
-        };
-        let sent_changes = value
-            .sent_changes
-            .into_iter()
-            .map(Change::from_bytes)
-            .collect::<Result<_, _>>()?;
-        Ok(Self {
-            shared_heads: value.shared_heads,
-            last_sent_heads: value.last_sent_heads,
-            their_heads: value.their_heads,
-            their_need: value.their_need,
-            our_need: value.our_need,
-            their_have: have,
-            sent_changes,
-        })
-    }
-}
-
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct RawSyncMessage {
-    pub heads: Vec<ChangeHash>,
-    pub need: Vec<ChangeHash>,
-    pub have: Vec<RawSyncHave>,
-    pub changes: Vec<Vec<u8>>,
-}
-
-impl TryFrom<SyncMessage> for RawSyncMessage {
-    type Error = AutomergeError;
-
-    fn try_from(value: SyncMessage) -> Result<Self, Self::Error> {
-        let have = value
-            .have
-            .into_iter()
-            .map(RawSyncHave::try_from)
-            .collect::<Result<_, _>>()?;
-        let changes = value
-            .changes
-            .into_iter()
-            .map(|c| c.raw_bytes().to_vec())
-            .collect();
-        Ok(Self {
-            heads: value.heads,
-            need: value.need,
-            have,
-            changes,
-        })
-    }
-}
-
-impl TryFrom<RawSyncMessage> for SyncMessage {
-    type Error = AutomergeError;
-
-    fn try_from(value: RawSyncMessage) -> Result<Self, Self::Error> {
-        let have = value
-            .have
-            .into_iter()
-            .map(SyncHave::try_from)
-            .collect::<Result<_, _>>()?;
-        let changes = value
-            .changes
-            .into_iter()
-            .map(Change::from_bytes)
-            .collect::<Result<_, _>>()?;
-        Ok(Self {
-            heads: value.heads,
-            need: value.need,
-            have,
-            changes,
-        })
-    }
-}
-
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct RawSyncHave {
-    pub last_sync: Vec<ChangeHash>,
-    #[serde(with = "serde_bytes")]
-    pub bloom: Vec<u8>,
-}
-
-impl TryFrom<SyncHave> for RawSyncHave {
-    type Error = AutomergeError;
-
-    fn try_from(value: SyncHave) -> Result<Self, Self::Error> {
-        Ok(Self {
-            last_sync: value.last_sync,
-            bloom: value.bloom.into_bytes()?,
-        })
-    }
-}
-
-impl TryFrom<RawSyncHave> for SyncHave {
-    type Error = AutomergeError;
-
-    fn try_from(raw: RawSyncHave) -> Result<Self, Self::Error> {
-        Ok(Self {
-            last_sync: raw.last_sync,
-            bloom: BloomFilter::try_from(raw.bloom.as_slice())?,
-        })
-    }
 }
 
 #[wasm_bindgen(js_name = generateSyncMessage)]
